fix: generate-commit and add commit-editor to the project
This commit adds the gencommits package to the project and creates a new script called generate:commit in the package.json file. This script can be used to generate a commit message based on the type of change that is being made.

The commit-editor package is also added to the project and is used to open a text editor when the generate:commit script is run. This allows the user to customize the commit message before it is saved.

The conventional commits standard is used to format the commit messages. This standard helps to keep the commit history clean and organized.

BREAKING CHANGE: The generate:commit script replaces the old commit message generation logic.